+++ coroutine/win64/Context.h
@@ -53,7 +53,7 @@ static inline void coroutine_initialize(
 
     /* Return address */
     *--context->stack_pointer = 0;
-    *--context->stack_pointer = (void*)start;
+    *--context->stack_pointer = (void*)(uintptr_t)start;
     *--context->stack_pointer = (void*)coroutine_trampoline;
 
     /* Windows Thread Information Block */
